requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.7
django-crispy-forms==2.0
django-environ==0.10.0
django-import-export==3.1.0
django-sendgrid-v5==1.2.2
django-storages==1.13.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
MarkupPy==1.14
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tablib==3.3.0
tomli==2.0.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
